+++ packages/lib/createTestMountParameters/index.ts
@@ -59,6 +59,9 @@ export default (): MockParameters => {
       errors: [],
       plugin: {},
       language: '',
+      zoomLevel: 0,
+      hovered: 0,
+      selected: 0,
     },
   }
 }