+++ packages/plugins/Gfi/src/components/FeatureTableHead.vue
@@ -11,7 +11,7 @@
 import Vue from 'vue'
 
 export default Vue.extend({
-  name: 'GfiHead',
+  name: 'GfiFeatureTableHead',
 })
 </script>