+++ packages/plugins/Gfi/src/types.ts
@@ -1,4 +1,4 @@
-import { Map } from 'ol'
+import { Map, Feature } from 'ol'
 import { Coordinate } from 'ol/coordinate'
 import { SourceType } from 'ol/layer/WebGLTile'
 import { LayerType } from 'ol/renderer/webgl/TileLayer'
@@ -10,6 +10,8 @@ import {
   GfiAfterLoadFunction,
   GfiConfiguration,
   GfiLayerConfiguration,
+  RenderType,
+  FeatureList,
 } from '@polar/lib-custom-types'
 
 /** parameter specification for request method */
@@ -32,6 +34,7 @@ export interface GfiState {
   visibleWindowFeatureIndex: number
   /** default style for stroke and fill of the highlighted feature. */
   defaultHighlightStyle: HighlightStyle
+  page: number
 }
 
 export interface GfiGetters extends GfiState {
@@ -41,6 +44,7 @@ export interface GfiGetters extends GfiState {
   gfiConfiguration: GfiConfiguration
   /** all layer keys to retrieve GFI information for */
   layerKeys: string[]
+  renderType: RenderType
   /** subset of layerKeys, where features' properties are to be shown in UI */
   windowLayerKeys: string[]
   /**
@@ -50,4 +54,9 @@ export interface GfiGetters extends GfiState {
   windowFeatures: GeoJsonProperties[]
   /** subset of layerKeys, where features' geometries are to be shown on map */
   geometryLayerKeys: string[]
+  windowLayerKeysActive: boolean
+  listMode: FeatureList['mode'] | undefined
+  listText: FeatureList['text']
+  showList: boolean
+  listFeatures: Feature[]
 }