+++ packages/plugins/IconMenu/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unpublished
 
+- Feature: Add default locale for gfi plugin as child.
 - Fix: Resolved internal error. (No effect on client behaviour.)
 
 ## 1.0.1