+++ packages/types/custom/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Feature: Currently active `language` key has been added to the core store.
 - Feature: Add types for new plugin `Filter`, namely `FilterConfiguration`; also to `MapConfig` as configurable plugin.
+- Feature: Add types for new core state and getters. (`zoomLevel`, `hovered`, `selected`)
+- Feature: Add parameters `renderType`, `featureList`, and `activeLayerPath` to `GfiConfiguration`.
+- Feature: Add type `FeatureList` for `GfiConfiguration`.
 - Feature: Add parameter `renderType` to `GeoLocationConfiguration`.
 - Feature: Add export of `RenderType`.
 - Feature: Add parameter `focusAfterSearch` to `AddressSearchConfiguration`.